% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{available_palettes}
\alias{available_palettes}
\title{Palette options
Generate a tibble of with palette options}
\usage{
available_palettes(team = "all")
}
\arguments{
\item{team}{return all available colors for a given team. Default is
all available teams}
}
\value{
A tibble containing the palette options for a given NBA team.
}
\description{
Palette options
Generate a tibble of with palette options
}
\examples{
available_palettes()#'
available_palettes('raptors')#'
available_palettes('heat')
}
\keyword{colors}
