% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nba_results2020}
\alias{nba_results2020}
\title{NBA game results for the 2019-2020 season.}
\format{
A data frame with 2118 rows and 47 variables:
}
\source{
\url{http://asbcllc.com/nbastatR/reference/game_logs.html}
}
\usage{
nba_results2020
}
\description{
A dataset containing results and total team statistics
from each regular season NBA game
from the 2019-2020 season. The data is can be obtained by
using the `game_logs()` function from the `nbastatR` package
}
\keyword{datasets}
